requirements.txt
aniso8601==3.0.2
argon2-cffi==19.1.0
cffi==1.11.5
click==6.7
coverage==4.5.1
Django==1.9.9
django-debug-toolbar==1.0
Flask==1.0.2
Flask-HTTPAuth==3.2.4
Flask-RESTful==0.3.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
peewee==3.7.0
Pillow==5.3.0
pycparser==2.19
pytz==2018.5
selenium==2.53.1
six==1.11.0
sqlparse==0.2.1
Werkzeug==0.14.1
